+++ functions.py
@@ -7,4 +7,16 @@ def minus(a, b):
     return a - b
 
 def share(a, b):
+  """ """"
     return a // b
+  
+def modulo(a, b):
+    """a mod b"""
+    if b == 0:
+        print("Can't operate!!")
+    else:
+        return a % b
+      
+def multiply(a, b):
+    """Multiply two numbers"""
+    return a * b