+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+----------------------
+
 ## [1.0.0] - 2025-05-11
 
 ### Added
@@ -12,4 +14,6 @@
 - Created `requirements.txt` for managing project dependencies.
 - Provided `README.md` with basic project information and setup instructions.
 
-- Added `share(a, b)` function in `functions.py` to perform multiplication.
+- Added share function in `functions.py`
+- Added modulo function in `functions.py`
+- Added multiply function in `functions.py`